+++ libathemecore/conf.c
@@ -231,6 +231,8 @@ void init_newconf(void)
 	add_bool_conf_item("LEAVE_CHANS", &conf_gi_table, 0, &config_options.leave_chans, false);
 	add_bool_conf_item("KLINE_WITH_IDENT", &conf_gi_table, 0, &config_options.kline_with_ident, false);
 	add_bool_conf_item("KLINE_VERIFIED_IDENT", &conf_gi_table, 0, &config_options.kline_verified_ident, false);
+	add_bool_conf_item("AKILL_LIST_AUTO_KLINES", &conf_gi_table, 0, &config_options.akill_list_auto_klines, false);
+	add_bool_conf_item("AKILL_LIST_MASS_KLINES", &conf_gi_table, 0, &config_options.akill_list_mass_klines, false);
 	add_conf_item("UFLAGS", &conf_gi_table, c_gi_uflags);
 	add_conf_item("CFLAGS", &conf_gi_table, c_gi_cflags);
 	add_bool_conf_item("RAW", &conf_gi_table, 0, &config_options.raw, false);