+++ modules/operserv/clearchan.c
@@ -120,7 +120,7 @@ static void os_cmd_clearchan(sourceinfo_t *si, int parc, char *parv[])
 						command_success_nodata(si, _("\2CLEARCHAN\2: Not klining exempt user %s!%s@%s"),
 								cu->user->nick, cu->user->user, cu->user->host);
 					else
-						kline_add_user(cu->user, reason, 604800, get_storage_oper_name(si));
+						kline_add_user(cu->user, reason, 604800, get_storage_oper_name(si), KLINE_MASS);
 			}
 		}
 	}